feat(core): add MT19937 PRNG, settings evaluator, stroke_to skeleton
M2 scaffolding for the stroke engine.

- rng: real MT19937 (replaces xorshift stub) with reference output test
  for seed 5489; next_unit / next_unit_f64 / next_gauss helpers
- evaluator: InputValues + SettingValues + evaluate(brush, inputs)
  implementing base_value + Σ mapping.eval(input_value)
- stroke: minimal stroke_to that builds inputs from the event, evaluates
  settings, accumulates fractional dabs from dabs_per_actual_radius /
  dabs_per_basic_radius / dabs_per_second, and emits dabs along the
  segment via TiledSurface::draw_dab
- state: BrushState seed widened/narrowed to u32 to match MT19937

Intentionally deferred (TODO(M2-followup)): slow_tracking, speed
slowness low-pass, tracking_noise, attack, offset_by_random/speed,
change_color_* drift, smudge bucket (needs M3 get_color),
stroke_threshold, tilt-derived inputs, recursive custom input.

+++ crates/hokusai-core/src/state.rs
@@ -49,7 +49,7 @@ pub struct BrushState {
 }
 
 impl BrushState {
-    pub fn new(seed: u64) -> Self {
+    pub fn new(seed: u32) -> Self {
         Self {
             actual_x: 0.0,
             actual_y: 0.0,
